[x64] make pjit_test compatible with strict dtype promotion

+++ tests/pjit_test.py
@@ -435,10 +435,10 @@ class PJitTest(jtu.BufferDonationTestCase):
   @jtu.with_mesh([('x', 2), ('y', 1)])
   def testNested(self):
     # Add a constant captured by the nested pjit to make things more complicated
-    h = jnp.arange(4)
+    h = jnp.arange(4.)
     f = pjit(lambda x: x.sum() + h.sum(), in_axis_resources=P('x', 'y'), out_axis_resources=None)
     g = pjit(lambda x: f(jnp.sin(x)), in_axis_resources=P('x', None), out_axis_resources=None)
-    x = jnp.arange(16).reshape((4, 4))
+    x = jnp.arange(16.).reshape((4, 4))
     y = g(x)
     self.assertAllClose(y, jnp.sin(x).sum() + h.sum())
     self.assertTrue(hasattr(y, "sharding_spec"))
@@ -448,17 +448,16 @@ class PJitTest(jtu.BufferDonationTestCase):
     if len(mesh) != 2: return
     assert resources == ('x', 'y')
     # Add a constant captured by the nested pjit to make things more complicated
-    h = jnp.arange(4)
+    h = jnp.arange(4.)
     f = pjit(lambda x: x.sum(1) * h.sum(),
              in_axis_resources=P('x', 'y'), out_axis_resources=P(('x', 'y')))
     g = pjit(lambda x: f(jnp.sin(x * 4 + 2)),
              in_axis_resources=P('x', None), out_axis_resources=P(('x', 'y')))
-    jtu.check_grads(g, (jnp.arange(16, dtype=jnp.float32).reshape((4, 4)) / 100,),
-                    order=2)
+    jtu.check_grads(g, (jnp.arange(16.).reshape((4, 4)) / 100,), order=2)
 
   @jtu.with_mesh([('x', 2), ('y', 1)])
   def testEvalJaxpr(self):
-    x, y = jnp.arange(4), jnp.arange(5)
+    x, y = jnp.arange(4.), jnp.arange(5.)
     f = pjit(lambda x, y: x.sum() + jnp.sin(y),
              in_axis_resources=(P('x'), P('y')),
              out_axis_resources=P('y'))
